rename: queryParams.ts -> walletAddressSchema.ts in lib/positions
The file only contained a Zod wallet address schema, not query-param
parsing logic. Renamed to walletAddressSchema.ts for accuracy and
updated the single import site in app/api/liquidations/route.ts.
